lib: Use unsigned long in ber_write_OID_String
This is what smb_strtoul returns, so use it.

+++ lib/util/asn1.c
@@ -276,7 +276,7 @@ bool asn1_write_BitString(struct asn1_data *data, const void *p, size_t length, 
 
 bool ber_write_OID_String(TALLOC_CTX *mem_ctx, DATA_BLOB *blob, const char *OID)
 {
-	unsigned int v, v2;
+	unsigned long int v, v2;
 	const char *p = (const char *)OID;
 	char *newp;
 	int i;
